FIX: inconsistent view of data with expire handler
In this fix I take a jab at handling logical race condition in data view.
A key may have "expired" but blocked due to another party attempting to update
the state of the key e.g. Setexp, Set, Getset, Getexp, Expire.

In certain cases data view after the aforementioned operation should cause the
handler to halt.  So the fix is simply looking at the jobId designated from
timer scheduler.

If the jobId matched then the key is not modified in between
the time it was modified to the time it expired.  If the jobId changed, either
an invalid one (jobId should be positive integer), or reassigned, then the
expire handler should quit its operation.